Add :test rake task for {syntax,lint,spec}
Add a rake task/alias to call the most common test tasks in the correct
order. Make TravisCI run this on all tests.

Unfortunately we can't define :default because it has been claimed by
puppetlabs_spec_helper to print out the list of available tasks. Or rather,
we can, but the output looks messy.