Add package resource and remove service provider
Remove the `provider` param from the `service{}` resource and the respective
variable from the `::params` class. I don't think this is a very good
demonstration of the params class. Not least because Puppet should do a
reasonable job of detecting the right provider for most platforms.

Replace this with `$package_name` and `$service_name` variables, because
these are things that are more likely to change on a per-distro basis. Added
a `package{}` resource accordingly.